change move slider to a drop-down box
mostly because the board updates were much too slow for the slider to be
comfortable; also <input type=range> is not super compatible yet.

The slowness could probably be addressed with a combination of cacheing
all board states in advance (I like persistent data structures such as
immutable.js or mori.js for this, but it's probably not necessary), and
limiting DOM updates to only apply to board points that are
changing (React.js is a good candidate to handle this for us; it could
be coded manually but I think we're reaching the point where a framework
should be considered anyway).