add logging to files
Casper will write a file `../casper_tests.log`, which is cleared and
refilled on each test run.  Meanwhile the Flask server will write some
barebones logging messages to `../test.log`, continuing the log on each
run and periodically rotating the current file into `../test.log.1` (and
so on) when it fills up.